+++ Changelog.md
@@ -6,6 +6,9 @@
    been added.
    * [Pull request 329](https://bitbucket.org/ignitionrobotics/ign-transport/pull-requests/329)
 
+1. Added log::Recorder::Topics that returns the set of added topics.
+   * [Pull request 328](https://bitbucket.org/ignitionrobotics/ign-transport/pull-requests/328)
+
 1. Added log::Recorder::Filename that returns the name of the log file.
    * [Pull request 327](https://bitbucket.org/ignitionrobotics/ign-transport/pull-requests/327)